Refactor database utilities package and update resolvers
Moved DatabaseConnector and TableNames from utilities to a new database package and updated all imports accordingly. Renamed TypeResolver methods to videoQualityModeResolver and voiceChannelBitrateResolver, and updated all usages to match.
