+++ src/main/java/database/DatabaseConnector.java
@@ -1,10 +1,12 @@
-package utilities;
+package database;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+
+import utilities.EnvConfig;
 
 public class DatabaseConnector {
