+++ src/main/java/listeners/auditloglistener/AuditLogCommandListener.java
@@ -3,13 +3,13 @@ package listeners.auditloglistener;
 import java.awt.Color;
 import java.sql.SQLException;
 
+import database.DatabaseConnector;
+import database.TableNames;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import utilities.DatabaseConnector;
-import utilities.TableNames;
 
 public class AuditLogCommandListener extends ListenerAdapter {
